requirements.txt
python-telegram-bot==20.7
httpx==0.25.2
websockets==12.0
base58==2.1.1
